# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "summary": "List my notifications (contract §4.12)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/notifications/read": {
      "post": {
        "operationId": "NotificationsController_markRead",
        "summary": "Mark notifications read (contract §4.12)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/devices": {
      "post": {
        "operationId": "NotificationsController_register",
        "summary": "Register a push device token (contract §4.12)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/devices/{token}": {
      "delete": {
        "operationId": "NotificationsController_remove",
        "summary": "Unregister a push device (contract §4.12)",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/posts": {
      "post": {
        "operationId": "PostsController_compose",
        "summary": "Compose a post (contract §4.6)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/posts/{id}": {
      "patch": {
        "operationId": "PostsController_update",
        "summary": "Edit own post (contract §5b)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PostsController_remove",
        "summary": "Delete own post (contract §5b)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/posts/{id}/like": {
      "post": {
        "operationId": "PostsController_like",
        "summary": "Like a post (contract §4.5)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikesDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PostsController_unlike",
        "summary": "Unlike a post (contract §4.5)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikesDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/posts/{id}/save": {
      "post": {
        "operationId": "PostsController_save",
        "summary": "Save a post (contract §4.5)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "PostsController_unsave",
        "summary": "Unsave a post (contract §4.5)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/posts/{id}/comments": {
      "get": {
        "operationId": "PostsController_comments",
        "summary": "List comments on a post (contract §4.5)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "PostsController_addComment",
        "summary": "Add a comment (contract §4.5)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        },
        "tags": [
          "posts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/copy/{traderId}/start": {
      "post": {
        "operationId": "CopyController_start",
        "summary": "Start copying a trader (contract §4.7)",
        "parameters": [
          {
            "name": "traderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartCopyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopyConfigDto"
                }
              }
            }
          }
        },
        "tags": [
          "copy"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/copy/trade/{postId}": {
      "post": {
        "operationId": "CopyController_copyTrade",
        "summary": "Copy a single posted trade from the wallet (contract §4.6)",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyTradeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CopyPositionDto"
                }
              }
            }
          }
        },
        "tags": [
          "copy"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/copy/{traderId}/stop": {
      "post": {
        "operationId": "CopyController_stop",
        "summary": "Stop copying a trader; closes open positions (contract §4.7)",
        "parameters": [
          {
            "name": "traderId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "copy"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/copy": {
      "get": {
        "operationId": "CopyController_configs",
        "summary": "My active copy configs (contract §4.7)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CopyConfigDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "copy"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/positions": {
      "get": {
        "operationId": "CopyController_positions",
        "summary": "My copied positions with live P/L (contract §4.7)",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "active",
                "closed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CopyPositionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "copy"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/portfolio/summary": {
      "get": {
        "operationId": "CopyController_summary",
        "summary": "Portfolio summary (contract §4.7)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSummaryDto"
                }
              }
            }
          }
        },
        "tags": [
          "copy"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Millimore API",
    "description": "Backend API for the Millimore platform (mobile app + admin dashboard). Single source of truth: docs/BACKEND_CONTRACT.md",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.millimore.app/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "n_1"
          },
          "type": {
            "type": "string",
            "example": "creator.status"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "object",
            "nullable": true
          },
          "data": {
            "type": "object",
            "description": "Arbitrary payload (traderId, pair, pnl…)."
          },
          "createdAt": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "createdAt",
          "read"
        ]
      },
      "MarkReadDto": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ]
          },
          "token": {
            "type": "string",
            "description": "FCM/APNs device token."
          },
          "appVersion": {
            "type": "string"
          }
        },
        "required": [
          "platform",
          "token"
        ]
      },
...
      "CreateTradeDto": {
        "type": "object",
        "properties": {
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ]
          },
          "entryType": {
            "type": "string",
            "enum": [
              "market",
              "limit"
            ]
          },
          "entryPrice": {
            "type": "number"
          },
          "sl": {
            "type": "number"
          },
          "tp": {
            "type": "number"
          },
          "slPct": {
            "type": "number"
          },
          "tpPct": {
            "type": "number"
          }
        }
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "trade",
              "analysis",
              "lesson",
              "update"
            ]
          },
          "content": {
            "type": "string"
          },
          "pair": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imageUrl": {
            "type": "string",
            "description": "Uploaded image URL (from POST /uploads)."
          },
          "trade": {
            "description": "Trade payload (type=trade) — makes the post copyable.",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateTradeDto"
              }
            ]
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ]
          },
          "entryType": {
            "type": "string",
            "enum": [
              "market",
              "limit"
            ]
          },
          "entryPrice": {
            "type": "number"
          },
          "sl": {
            "type": "number"
          },
          "tp": {
            "type": "number"
          },
          "slPct": {
            "type": "number"
          },
          "tpPct": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "content"
        ]
      },
...
      "PostDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "p_1"
          },
          "trader": {
            "$ref": "#/components/schemas/TraderDto"
          },
          "type": {
            "type": "string",
            "enum": [
              "analysis",
              "trade",
              "lesson",
              "update"
            ]
          },
          "content": {
            "type": "string"
          },
          "pair": {
            "type": "object",
            "nullable": true
          },
          "title": {
            "type": "object",
            "nullable": true
          },
          "points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "imageUrl": {
            "type": "object",
            "nullable": true,
            "description": "Chart/setup image."
          },
          "trade": {
            "nullable": true,
            "description": "Present on trade posts.",
            "allOf": [
              {
                "$ref": "#/components/schemas/TradePostDto"
              }
            ]
          },
          "likes": {
            "type": "number",
            "example": 284
          },
          "comments": {
            "type": "number",
            "example": 47
          },
          "createdAt": {
            "type": "string"
          },
          "isLiked": {
            "type": "boolean"
          },
          "saved": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "trader",
          "type",
          "content",
          "points",
          "likes",
          "comments",
          "createdAt",
          "isLiked",
          "saved"
        ]
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "pair": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "LikesDto": {
        "type": "object",
        "properties": {
          "likes": {
            "type": "number",
            "example": 285
          }
        },
        "required": [
          "likes"
        ]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "c_1"
          },
          "author": {
            "type": "string",
            "example": "Priya"
          },
          "username": {
            "type": "string",
            "example": "priyatrades"
          },
          "text": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "byMe": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "author",
          "username",
          "text",
          "createdAt",
          "byMe"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      },
...
      "StartCopyDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "example": "acc_1",
            "description": "Optional broker account hint. Funding comes from the wallet."
          },
          "amount": {
            "type": "number",
            "example": 500,
            "description": "Margin to allocate from the wallet."
          },
          "leverage": {
            "type": "number",
            "example": 100,
            "description": "Leverage (1..maxLeverage). Defaults to your setting."
          },
          "risk": {
            "type": "number",
            "example": 1,
            "default": 1
          },
          "autoCopy": {
            "type": "boolean",
            "example": true,
            "default": true
          }
        },
        "required": [
          "amount"
        ]
      },
      "CopyConfigDto": {
        "type": "object",
        "properties": {
          "traderId": {
            "type": "string",
            "example": "t_1"
          },
          "accountId": {
            "type": "string",
            "example": "acc_1"
          },
          "amount": {
            "type": "number",
            "example": 500,
            "description": "Margin allocated from the wallet."
          },
          "leverage": {
            "type": "number",
            "example": 100,
            "description": "Leverage; exposure = amount × leverage."
          },
          "risk": {
            "type": "number",
            "example": 1
          },
          "autoCopy": {
            "type": "boolean",
            "example": true
          },
          "startedAt": {
            "type": "string"
          }
        },
        "required": [
          "traderId",
          "accountId",
          "amount",
          "leverage",
          "risk",
          "autoCopy",
          "startedAt"
        ]
      },
      "CopyTradeDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 200,
            "description": "Margin to allocate from the wallet."
          },
          "leverage": {
            "type": "number",
            "example": 100,
            "description": "Leverage (1..maxLeverage). Defaults to your setting."
          }
        },
        "required": [
          "amount"
        ]
      },
      "CopyPositionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pos_1"
          },
          "traderId": {
            "type": "string",
            "example": "t_1"
          },
          "traderName": {
            "type": "string",
            "example": "Marcus"
          },
          "pair": {
            "type": "string",
            "example": "EUR/USD"
          },
          "isBuy": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "closed"
            ]
          },
          "entryPrice": {
            "type": "number",
            "example": 1.0876
          },
          "exitPrice": {
            "type": "object",
            "nullable": true
          },
          "pnlAmount": {
            "type": "number",
            "example": 12.4
          },
          "pnlPercent": {
            "type": "number",
            "example": 0.47
          },
          "lots": {
            "type": "number",
            "example": 0.1
          },
          "openedAt": {
            "type": "string"
          },
          "closedAt": {
            "type": "object",
            "nullable": true
          },
          "accountId": {
            "type": "string",
            "example": "acc_1"
          }
        },
        "required": [
          "id",
          "traderId",
          "traderName",
          "pair",
          "isBuy",
          "status",
          "entryPrice",
          "pnlAmount",
          "pnlPercent",
          "lots",
          "openedAt",
          "accountId"
        ]
      },
...
    }
  }
}